+++ .luarc.json
@@ -1,9 +1,10 @@
 {
     "$schema": "https://raw.githubusercontent.com/sumneko/vscode-lua/master/setting/schema.json",
-    "Lua.diagnostics.disable": [
+    "runtime.version": "Lua 5.1",
+    "diagnostics.disable": [
         "ambiguity-1"
     ],
-    "Lua.diagnostics.globals": [
+    "diagnostics.globals": [
         "xero"
     ]
 }
