+++ template/mirin/options.lua
@@ -1,0 +1,7 @@
+-- Convenience shortcuts / options
+
+return {
+    max_pn = 8, -- default: `8`
+    debug_print_applymodifier_input = false, -- default: `false`
+    debug_print_mod_targets = false, -- default: `false`
+}
